pull all members from Rover class to abstract class VehicleBase
